basic setup for download all clinical data
